Move showcase screen out of package
