+++ sample-app/src/main/kotlin/com/sats/dna/sample/components/ChipsScreen.kt
@@ -1,4 +1,4 @@
-package com.sats.dna.sample.components.chip
+package com.sats.dna.sample.components
 
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
